Updat the code

+++ python-serialization/task_02_csv.py
@@ -3,9 +3,9 @@
 This module provides a function that convert
 csv data into json and save it on the file
 """
-
 import csv
 import json
+import os
 
 
 def convert_csv_to_json(csv_filename):
@@ -13,9 +13,18 @@ def convert_csv_to_json(csv_filename):
     This function convert the content of csv file into json
     and save it on the another file
     """
-    with open(csv_filename, 'r') as f:
-        reader = csv.DictReader(f)
-        data = [row for row in reader]
 
-    with open('data.json', 'w', encoding='utf8') as f:
-        json.dump(data, f, indent=4)
+    try:
+        if not os.path.exists(csv_filename):
+            raise FileNotFoundError("File not found")
+
+        with open(csv_filename, 'r') as f:
+            reader = csv.DictReader(f)
+            data = [row for row in reader]
+
+        with open('data.json', 'w', encoding='utf8') as f:
+            json.dump(data, f, indent=4)
+
+    except FileNotFoundError as e:
+        print(e)
+        return False
